firewall: Add explicit feature API
Instead heaving a generic firewall API, we use an just provide
a few features such as NAT, SNAT or MARK. That allows
us to push down all the code into firewall.c.

There are only two different users in ConnMan for this API:
enabling global NAT or for session handling.

Fortunately, the NAT handling is pretty simple. We just have one
global table and we either have it enabled or disabled.

The session handling is slightly more tricky. There are
three different rule sets. The first one enables one global rule
for CONNMARK save/restore. This one will be installed only when the
first session calls __connman_firewall_enable_marking(). For this we
use a global context and enable disable it only if there is a session
using it.

The SNAT rules are shared between sessions and there exists only one
per output device. The session tracking is done in session.c. Note
the SNAT rules have also their own struct firewall_context.

The third set of rule for MARK is owned by the session, that is the
rules are tracked it the session->fw context.

Due to the separation of the struct firewall_context usage we can
git rid of  any special tracking. Either we enable all or disable all
in one go. We don't have enable indiviual rules anymore with the
FW_ALL_RULES. And we are all happy again.