dhcp: Follow the RFC regarding ordrer of "subnet" and "router" option
The 2132 RFC clearly state, that the "subnet" option should be before
the "router" option.

"If both the subnet mask and the router option are specified in a DHCP
reply, the subnet mask option MUST be first."
- https://tools.ietf.org/html/rfc2132#section-3.3

Some dhcp servers (D-Link DSR-1000) doesn't responds correctly, if that
isn't the case.
For example the D-Link DSR-1000 which respond back with a DHCPACK
containing two "subnet" option but no "router" option.