rootnfs: Working rootnfs using connman
Until now for root NFS you either had to manually blacklist
the interface or disable connman all together

This patch automatically blacklists the interface the NFS server
is reachable from and populates the resolver entries that the
DHCP server provided on startup.

It is now possible to use a vanilla rootfs tarball without
having to manually edit connman configuration entries.