connman.service.in: Drop CAP_KILL from upstream service file
Upstream ConnMan uses only D-Bus to start external programs. The
CAP_KILL capability in the service file is not necessary and can
therefore be removed. ConnMan still has the capability of starting
other applications, if this is needed the capability needs to be
re-added to the distro or OS specific service file.