gsupplicant: Use ConfigFile when recreating an interface
Currently, each time ConnMan removes a wpa_s interface the configuration of the
wpa_s for that interface is lost (See use cases below), such configuration is
present in the file passed to wpa_s using "-c" option. It is due to wpa_s
did not provide the configuration file path to the applications on top of it
(ConnMan in this case) in order to allow them to recreate interfaces with the
same configuration they were initially created.

In order to solve this we proposed a patch to wpa_s community, that patch aimed
to expose the configuration file path as a D-Bus interface property. It was
applied (b44d9c7) and delivered on last wpa_s release (2.6). Now, ConnMan has
the possibility to recreate interfaces with the correct configuration file path
in order to solve the described issue.

Two use cases are the following:

1. Enable/Disable WiFi: All wpa_s interfaces are removed then config information
is lost for all of them.
2. Tethering: When enabling, the selected wpa_s interface is removed and
recreated with the bridge "tether". Also when disabling, again the tethering
wpa_s interface is removed and recreated without bridge. After
enabling/disabling tethering the config information of that interface is lost.

This patch aims to solve the just described issue by using the new interface's
property exposed by wpa_s via D-Bus in order to always recreate the interfaces
with the appropriate configuration. This modification only has effect if wpa_s
is updated to release 2.6 or at least includes commit b44d9c7 (Applied on
2016-08-23), otherwise ConnMan will simply do not send any ConfigFile while
creating a new interface as it has been doing so far.