ofono: Fix segfault during set_property
If the SIM card is ejected or the modem is reset / removed between the
set_property call and the set_property_reply callback, connmand will generally
abort, as the remove_all_contexts will get called, and the callback accesses a
reference to a stale context via the property_info pointer.

The following patch adds reference counting to the network_context struct to
prevent the segfault.